Make Computer generic over its memory type

+++ src/sec1/day02.rs
@@ -1,20 +1,20 @@
 use crate::comp::Computer;
 use std::str::FromStr;
 
-pub fn run_with_args(c: &mut Computer, noun: isize, verb: isize) -> isize {
+pub fn run_with_args(c: &mut Computer, noun: i32, verb: i32) -> i32 {
     c.abs_store(1, noun);
     c.abs_store(2, verb);
     c.run().abs_load(0)
 }
 
 #[aoc(day2, part1)]
-pub fn p1(input: &str) -> isize {
+pub fn p1(input: &str) -> i32 {
     let mut c = Computer::from_str(input).unwrap();
     run_with_args(&mut c, 12, 2)
 }
 
 #[aoc(day2, part2)]
-pub fn p2(input: &str) -> isize {
+pub fn p2(input: &str) -> i32 {
     let mut c = Computer::from_str(input).unwrap();
     for n in 0..100 {
         for v in 0..100 {
